feat(cli): add CreateDatabaseToken function to generate a token for accessing the database
The CreateDatabaseToken function is added to generate a token for
accessing the database. It uses the provided email and password to
obtain a token from the server. The email flag is added to the
CreateDatabaseTokenFlag function to allow the user to specify the email
of the user for which the token should be generated.