feat(main.go): add server flag to specify the address of the API server
The main.go file has been modified to include a new flag called
"server". This flag allows the user to specify the address of the API
server when running the baserow command line application. The flag is
required and can be specified using the "-s" alias. This change provides
more flexibility and customization options for managing the baserow
instance.