fix(client.go): add support for database token authentication in API requests
The changes in `client.go` add support for database token authentication
in API requests. If the `ContextDatabaseToken` value is present in the
context, it will be used to add the appropriate authorization header to
the request. In `configuration.go`, the `ContextDatabaseToken` constant
is added to support the new database token authentication. This constant
can be used as a key in the context to pass the database token for
authentication.