chore(build.yaml): refactor build workflow to use a reusable docker build action
The build workflow has been refactored to use a reusable docker build
action from the dictyBase/workflows repository. This improves
maintainability and reduces duplication of code across different
workflows. The new action is called "call-docker-build" and it depends
on the "test" job. It inherits secrets from the parent workflow and
passes the repository and ref information to the action.