refactor(client.go): add CreateUser method to Client struct to create a new user
The CreateUser method is added to the Client struct to allow the
creation of a new user. It takes a token string and a pointer to an
APIUsersPostReq struct as parameters. It sends a POST request to the
"/api/users" endpoint with the user data in JSON format. The response is
then decoded and the user ID is returned. This new method provides a
convenient way to create users in the client application.