chore(go.mod): add indirect dependency on github.com/jellydator/ttlcache/v3 v3.1.0
The go.mod file has been updated to include two new indirect
dependencies. The first one is github.com/jellydator/ttlcache/v3 v3.1.0,
which is added to the list of required modules. The second one is
golang.org/x/sync v0.1.0, which is also added as an indirect dependency.
These changes were made to ensure that the project has the necessary
dependencies for its functionality.