feat(registry.go): add support for TTL cache in the registry
The registry.go file has been modified to add support for a TTL cache.
The import statement for the ttlcache package has been added. The
constant TTLCacheKey has been defined to store the key for the TTL cache
in the registry.
Two new functions, SetTTLCache and GetTTLCache, have been added to the
registry. SetTTLCache allows setting the TTL cache in the registry, and
GetTTLCache allows retrieving the TTL cache from the registry. These
functions provide convenient access to the TTL cache functionality in
the registry.