feat(logto): add support for TTL cache in the logto command
The logto command now supports a TTL cache for storing temporary data.
This is achieved by adding the `github.com/jellydator/ttlcache/v3`
package as an import and creating a new TTL cache instance in the `main`
function. The TTL cache instance is then set in the registry using the
`SetTTLCache` function.
In the `registry` package, the `SetTTLCache` function has been fixed to
use the correct type for the `tcache` parameter, which is
`*ttlcache.Cache[string, int]`. This ensures that the TTL cache is
correctly stored in the registry.