feat(main.go): add import-user command with input flag and before hook to set up ttlcache and reader
The code related to initializing and setting the ttlcache has been
removed as it is no longer needed. The import-user command has been
added with an input flag to specify the input csv file. A before hook
has been added to set up the ttlcache and reader before executing the
import-user command.