feat(client.go): add function to add custom user information
This commit adds a new function `AddCustomUserInformation` to the
`Client` struct in the `client.go` file. This function allows the client
to add custom information to a user by sending a PATCH request to the
appropriate API endpoint. The function takes in the user's token, user
ID, and the custom data to be added. It converts the custom data to
JSON, creates a PATCH request with the necessary headers and content,
and sends the request to the API. Any errors encountered during the
process are handled and returned as an error.