feat(user.go): add support for importing user custom data from CSV file
The code now includes logic to import user custom data from a CSV file.
The `ImportUser` function has been updated to handle the import process.
It now parses the CSV file and creates a new user with the provided
information. Additionally, it sets the `isSubscribed` flag based on the
value in the CSV file and adds the custom data to the user using the
`AddCustomUserInformation` method from the `lclient` client.
If any errors occur during the import process, the function will exit
with an error message. Debug logs are also added to provide information
about the creation of the custom data for each user.