style(plasmid_inventory.go): improve readability by breaking long function calls into multiple lines
The changes were made to improve the readability of the code. Long
function calls and declarations were broken down into multiple lines.
This makes the code easier to read and understand, especially for
developers who are new to the project or are not familiar with the
specific section of the code.