feat(client.go): add new client.go file with SetS3Client and SetContentAPIClient functions
The new client.go file includes two functions: SetS3Client and
SetContentAPIClient. SetS3Client function is used to create a new S3
client and store it in the registry. This allows the application to
interact with S3 storage. The SetContentAPIClient function is used to
establish a connection with the content grpc api server and store the
client in the registry. This allows the application to interact with the
content grpc api server. Both functions improve the modularity and
organization of the code.