feat(action.go): add LoadContent function to handle S3 objects
The LoadContent function has been added to the action.go file. This
function is responsible for listing and reading objects from an S3
bucket. It uses the S3 client from the registry to list objects in the
specified S3 bucket and path. For each object, it retrieves the object's
information and logs the key of the object. This function will be useful
for handling and processing data stored in S3 buckets.