refactor(plasmid.go): replace ioutil with os and fs for directory reading
The ioutil package is being phased out as of Go 1.16, and its
functionality has been provisioned in other packages like os and fs.
This change replaces ioutil with os and fs for directory reading in the
NewPlasmidSeqReader function. Additionally, error handling has been
added for file info retrieval to ensure the program doesn't crash when
it encounters an error while retrieving file information.