chore(lint.yml): update Go version from 1.20 to 1.21 and linter version from v1.52.2 to v1.55.2
The Go version and linter version have been updated to keep the project
up-to-date with the latest tools and features. This ensures that the
project is using the most recent improvements and bug fixes from these
tools, which can help to improve the quality of the code and reduce
potential issues.