feat(action.go): implement namespace mapping for better readability and maintainability
A new function namespaceMap() has been added to map short namespace
codes to their full names. This improves readability and maintainability
of the code by replacing cryptic short codes with meaningful names. The
nameAndNamespace() function has been updated to use this new mapping,
which will make the code easier to understand and modify in the future.