feat(load.go): add updateTermRow function to update term row
The properties for adding and checking the existence of a term row were
very similar, so they have been merged into a single struct,
termRowProperties, to reduce redundancy and improve code
maintainability. A new function, updateTermRow, has been added to allow updating of term
rows. This provides more flexibility in managing term rows, as they can
now be updated as well as added and checked for existence.