refactor(load.go): update ontologyRow struct to improve data representation
The ontologyRow struct has been updated to better represent the data it
holds. The 'Order' field was removed as it was not necessary. The
'TermId' field was renamed from 'Id' to 'term_id' to avoid confusion
with the 'Id' field. A new field 'Name' was added to store the name of
the ontology term. The 'IsObsolete' field was renamed from 'Is_obsolete'
to 'is_obsolete' to maintain consistency in naming conventions.