refactor(ontology.go): simplify OntologyTableManager struct and remove unused functions
The OntologyTableManager struct has been simplified by embedding the
TableManager struct, removing the need to duplicate fields. The
CreateTable function has been removed as it was not being used. The
fieldInformation and toFieldNames functions have also been removed as
they were not being used. The fieldDefs function has been renamed to
FieldDefs and made public as it is now being used in other parts of the
code.