refactor(ontology.go): rename RemoveExtraField to RemoveField and add field parameter
The function RemoveExtraField was renamed to RemoveField and a new
parameter 'field' was added to make the function more generic and
reusable. Now, it can remove any field from the table, not just the
'Field 1'. The hasExtraField function was removed as it is no longer
needed because the field to be removed is now passed as a parameter to
the RemoveField function.