refactor(ontology.go, table.go): move RemoveField function from ontology.go to table.go
The RemoveField function was moved from ontology.go to table.go. This
change was made to improve the organization of the codebase. The
function is more related to table operations, so it makes more sense to
have it in the table.go file. This will make the code easier to
understand and maintain.