refactor(table.go): modify CreateTable function to use HTTP request instead of API client
The CreateTable function was refactored to use an HTTP request instead
of the API client for table creation. This change was made to improve
the flexibility and control over the request. A new function,
CreteTableURL, was added to generate the URL for table creation. This
function improves code readability and maintainability. In the
CreateFields function, the variable name was changed from 'payload' to
'params' to better reflect its purpose and improve code readability.