feat(action.go): add option to save refresh token to a file
The changes allow the user to save the refresh token to a file. This is
useful for automating tasks that require a refresh token, as the token
can be read from the file instead of manually inputting it each time.
The file path can be specified using the "refresh-token-path" argument.
An error message is returned if there's an issue writing the token to
the file. A log message is also added to confirm the successful saving
of the refresh token.