fix(lint.yml): change ref from github.ref_name to github.event.pull_request.head.sha for accurate linting
The reference in the linting workflow has been changed from the branch
name to the specific commit SHA of the pull request head. This ensures
that the linting process is run on the exact code that is being proposed
for merging, rather than the latest code in the branch, which might
include changes that are not part of the pull request.