feat(ontology.go): add TabledateParams function to manage table date parameters
The TabledateParams function was added to manage the parameters related
to dates in the table. This includes parameters for 'is_obsolete',
'last_modified', and 'created_on'. This function will help in better
management and manipulation of date-related data in the table, providing
more control over how date and time information is handled and
displayed.