refactor(action.go): replace hardcoded field update with dynamic field updates
The hardcoded field update for "is_obsolete" has been replaced with a
loop that iterates over all the fields in the FieldChangeSpecs map. This
change allows for more flexibility and scalability, as it enables the
updating of any field specified in the FieldChangeSpecs map, rather than
just the "is_obsolete" field. This is particularly useful when there are
multiple fields that need to be updated.