feat(database): add PhenotypeTableManager struct and FieldNames method
A new file, phenotype.go, has been added to the database package. This
file introduces a new struct, PhenotypeTableManager, which embeds the
TableManager struct. A new method, FieldNames, has been added to the
PhenotypeTableManager struct. This method returns a slice of strings
representing the field names in the Phenotype table. This addition will
help in managing the Phenotype table and its fields more efficiently.